warnings: adapt to changes in QVariant type() / typeId()
In Qt5, type() returns a QVariant::Type, but the documentation says
to interpret this as a QMetaType::Type enumeration value. In Qt6,
this is an int, to be interpreted as QMetaType::Type.

Do so in both versions, and adapt the switch-case to follow.
The QVariant::String and other enum-values are documented as "obsolete"
even in Qt 5.15, and map to QMetaType::Type values.